Extracting logic, WIP
